Debugged eda script and added signal_selection.py
