+++ .vscode/launch.json
@@ -4,12 +4,18 @@
   // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
   "version": "0.2.0",
   "configurations": [
-
+    {
+      "name": "Python Debugger: Current File (no config)",
+      "type": "debugpy",
+      "request": "launch",
+      "program": "${file}",
+      "console": "integratedTerminal"
+    },
     {
       "name": "Python Debugger: train.py with denoising_config_debug.yaml",
       "type": "debugpy",
       "request": "launch",
-      "program": "code/denoising/train.py",
+      "program": "mycode/denoising/train.py",
       "console": "integratedTerminal",
       "args": ["--config", "code/denoising/configs/denoising_config_debug.yaml"]
     },
@@ -20,6 +26,13 @@
       "program": "tests/normalization/test_normalization.py",
       "console": "integratedTerminal",
       "args": ["--config", "config_channel.yaml"]
+    },
+    {
+      "name": "Python Debugger: ptbxl_eda.py without config",
+      "type": "debugpy",
+      "request": "launch",
+      "program": "mycode/eda/ptbxl_eda.py",
+      "console": "integratedTerminal",
     }
   ]
 }
