+++ code/eda/__init__.py
@@ -1,0 +1,7 @@
+"""
+EDA module for PTB-XL dataset analysis
+"""
+
+from . import eda_utils
+
+__all__ = ['eda_utils']
